Split up login form
Make log out only visible when logged in
